backend/requirements.txt
django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
requests==2.31.0
pillow==9.5.0